place redoc handler into separate crate
