+++ redoc/src/script.ts
