+++ redoc/src/webpack.config.ts
